Привнес свое в проект
1. Заголовок для картинки
2. Количество просомтров картинки
3. Новая миграция в БД
4. Заполнение поля updated_at при апдейте записи
5. Работа с условными операторами в шаблонизаторе

+++ app/Services/ImageServices.php
@@ -22,10 +22,10 @@ class ImageServices
         return $myImages;
     }
 
-    public function add($image)
+    public function add($image, $title)
     {
         $filename = $image->store('uploads');
-        DB::table('images')->insert(['image' => $filename]);
+        DB::table('images')->insert(['image' => $filename, 'title' => $title]);
 
         return true;
     }
@@ -40,10 +40,17 @@ class ImageServices
 
     public function update($id, $newImage)
     {
-        $image = $this->one($id);
-        Storage::delete($image->image);
-        $filename = $newImage->store('uploads');
-        DB::table('images')->where('id', $id)->update(['image' => $filename]);
+        $newIm = $newImage->image;
+        $newTitle = $newImage->title;
+        $updArr =[];
+        if ($newIm) {
+            $image = $this->one($id);
+            Storage::delete($image->image);
+            $filename = $newIm->store('uploads');
+            $updArr = ['image' => $filename];
+        }
+        $updArr = array_merge($updArr,['title' => $newTitle,'updated_at' => date("Y-m-d H:i:s")]);
+        DB::table('images')->where('id', $id)->update($updArr);
     }
 
     public function delete($id)
@@ -51,5 +58,15 @@ class ImageServices
         $image = $this->one($id);
         Storage::delete($image->image);
         DB::table('images')->where('id', $id)->delete();
+    }
+
+    /**
+     * Обновляем поле views при просмотре картинки
+     * @param $id
+     */
+    public function updateViews($id)
+    {
+        $viewsCount = DB::table('images')->select('views')->where('id', $id)->first();
+        DB::table('images')->where('id', $id)->update(['views' => $viewsCount->views+1]);
     }
 }
